backend/pyproject.toml
[project]
name = "reviewtalk-backend"
version = "0.1.0"
description = "AI 기반 제품 리뷰 챗봇 백엔드"
requires-python = ">=3.11,<3.12"

dependencies = [
    "fastapi>=0.110,<1.0",
    "uvicorn[standard]>=0.29,<1.0",
    "asyncpg>=0.29,<1.0",
    "sqlalchemy>=2.0,<3.0",
    "supabase>=2.3,<3.0",
    "pgvector>=0.2,<1.0",
    "langchain>=0.1,<1.0",
    "httpx>=0.27,<1.0",
    "beautifulsoup4>=4.12,<5.0",
    "openai>=1.25,<2.0",
    "pydantic>=2.6,<3.0",
    "python-dotenv>=1.0,<2.0",
    "tqdm>=4.66,<5.0",
    "loguru>=0.7,<1.0",
    "playwright>=1.43,<2.0",
    "google-generativeai>=0.7.0,<1.0.0"
]  